Extract Formatter out of TextUi, in accordance with SRP
